refactor(#82): 공통 타입/스키마를 shared 모듈로 이동 및 tsconfig 업그레이드
- ApiResponse 타입을 shard/types.ts로 이동
- ApiResponseSchema를 shared/schemas.ts로 이동
- 모든 도메인의 import 경로를 shared 모듈로 변경하여 의존성 제거
- tsconfig lib을 es2017에서 es2021로 업그레이드
