+++ packages/api/src/domains/directory/schemas.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 // User API에서 가져온 공통 응답 스키마
-import { ApiResponseSchema } from '../user/schemas';
+import { ApiResponseSchema } from '../../shared/schemas';
 
 // Directory 정보 스키마 (순환 참조 해결을 위해 lazy 사용)
 type DirectoryType = {
