+++ packages/api/src/shared/schemas.ts
@@ -1,1 +1,9 @@
-// 공통 스키마들은 실제 명세서 확인 후 추가
+import { z } from 'zod';
+
+// 공통 API 응답 스키마
+export const ApiResponseSchema = <T extends z.ZodTypeAny>(dataSchema: T) =>
+  z.object({
+    code: z.string(),
+    message: z.string(),
+    data: dataSchema,
+  });
