+++ packages/api/src/shared/types.ts
@@ -1,3 +1,10 @@
-// 공통 유틸리티 타입들만 남김
+// 공통 API 응답 타입
+export interface ApiResponse<T> {
+  code: string;
+  message: string;
+  data: T;
+}
+
+// 공통 유틸리티 타입들
 export type CreateRequest<T> = Omit<T, 'id' | 'createdAt' | 'updatedAt'>;
 export type UpdateRequest<T> = Partial<CreateRequest<T>>;
