+++ packages/api/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "./dist",
     "rootDir": "./src",
     "incremental": false,
-    "lib": ["dom", "dom.iterable", "es2017"]
+    "lib": ["dom", "dom.iterable", "es2021"]
   },
   "include": ["src"],
   "exclude": ["dist", "node_modules"]
